backend/pyproject.toml
[project]
name = "replaybench-backend"
version = "0.1.0"
description = "ReplayBench backend — HUD eval runner and trace ingestion pipeline"
requires-python = ">=3.11"
dependencies = [
    # Browser agent runtime
    "browser-use>=0.1.40",
    "playwright>=1.49.0",
    # LLM providers
    "openai>=1.50.0",
    "anthropic>=0.37.0",
    "google-genai>=1.0.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "langchain-google-genai>=2.0.0",
    # Observability
    "lmnr>=0.4.30",
    # Agent evaluation
    "hud-python>=0.5.26",
    # API server
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    # Backend communication
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.9.0",
    # MongoDB async driver (raw trace archive)
    "motor>=3.3.0",
    # Utilities
    "pyyaml>=6.0",
    "asyncio-throttle>=1.0.0",
]

[dependency-groups]
dev = []
